requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python-headless==4.8.1.78
numpy==1.24.3
pillow==10.0.1
ultralytics==8.0.196
tensorflow-cpu==2.13.0
python-dotenv==1.0.0
langchain==0.0.335
groq==0.4.1
langchain-groq==0.0.1
langchain-community==0.0.8
langchain-core==0.1.3
gunicorn==21.2.0